+++ Changelog.md
@@ -14,10 +14,12 @@
   - Alert
   - Badge
   - Floating Label
+  - CancelButton, SaveButton, CreateAction, StandardActions
 - Added custom `tag` support for card component
 - Added the following icons: `upload`,
 - Added the feature to substitute array values and/or object properties of any depth in widgets
 - Added the feature to use clojures in widget color definitions
+- Added the `text_if_null` widget modifier
 - Dropped Laravel 9 support
 - Dropped Enum v3 support
 - Replaced the multiselect dropdown with a nice-select2 implementation